Move the settings view files to the correct package
